refactor: Moved backend fetch logic to api/fetchOpenings.js for cleaner App.js
